requirements.txt
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.3
matplotlib>=3.7.1
pillow>=10.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
seaborn>=0.12.2
pandas>=2.0.3
tqdm>=4.65.0
gradio>=3.50.0
albumentations>=1.3.1
